Main changes:
- Godot 3 and 4 textures (most encodings)
  - Including ASTC for Sonic Colors Ultimate
- ASTC codec and container formats
  - Using the ASTC decoder from Basis Universal.
- Various format warnings are now errors.
- Fix DXT3 palette selection.
- Disable ZLIB_DEBUG in debug builds.
- Fix various issues with DXTn, ETCn, and PVRTC image sizes.
- New rp_image::swapRB() function to swap R and B channels, including
  an SSSE3-optimized version.
- Image rescaling for e.g. textures that have a power-of-2 size but
  should be rescaled to a smaller size for display purposes.

+++ src/kde/RomThumbCreator.cpp
@@ -2,7 +2,7 @@
  * ROM Properties Page shell extension. (KDE4/KF5)                         *
  * RomThumbCreator.cpp: Thumbnail creator.                                 *
  *                                                                         *
- * Copyright (c) 2016-2020 by David Korth.                                 *
+ * Copyright (c) 2016-2021 by David Korth.                                 *
  * SPDX-License-Identifier: GPL-2.0-or-later                               *
  ***************************************************************************/
 
@@ -275,7 +275,8 @@ Q_DECL_EXPORT int RP_C_API rp_create_thumbnail(const char *source_file, const ch
 	// in order to return better error codes.
 
 	if (getuid() == 0 || geteuid() == 0) {
-		qCritical("*** " RP_KDE_LOWER "%u does not support running as root.", QT_VERSION >> 16);
+		qCritical("*** " RP_KDE_LOWER "%u does not support running as root.",
+			static_cast<unsigned int>(QT_VERSION) >> 16);
 		return RPCT_RUNNING_AS_ROOT;
 	}
